Rename white_space_tokenizer to whitespace_tokenizer
And add it to registry under "whitespace".
