+++ allennlp/data/tokenizers/whitespace_tokenizer.py
@@ -6,13 +6,13 @@ from allennlp.data.tokenizers.token import Token
 from allennlp.data.tokenizers.tokenizer import Tokenizer
 
 
+@Tokenizer.register("whitespace")
 @Tokenizer.register("just_spaces")
 class WhitespaceTokenizer(Tokenizer):
     """
     A ``Tokenizer`` that assumes you've already done your own tokenization somehow and have
     separated the tokens by spaces.  We just split the input string on whitespace and return the
-    resulting list.  We use a somewhat odd name here to avoid coming too close to the more
-    commonly used ``SpacyTokenizer``.
+    resulting list.
 
     Note that we use ``text.split()``, which means that the amount of whitespace between the
     tokens does not matter.  This will never result in spaces being included as tokens.
